xamppp/htdocs/project/apps-mahalanobis/requirements.txt
absl-py==0.9.0
algorithmia==1.2.1
algorithmia-api-client==1.1.0
apyori==1.1.1
arff==0.9
astor==0.8.1
astroid==2.2.5
attrs==19.1.0
backcall==0.1.0
beautifulsoup4==4.8.1
bleach==1.5.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.1
Counter==1.0.0
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
Django==2.2.4
entrypoints==0.3
enum34==1.1.8.
falcon==2.0.0
gast==0.2.2
google-auth==1.10.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
graphviz==0.10.1
grpcio==1.26.0
h5py==2.10.0
html5lib==0.9999999
hug==2.5.4
idna==2.8
imageio==2.5.0
imutils==0.5.2
ipykernel==5.1.0
ipython==7.4.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.19
jedi==0.13.3
Jinja2==2.10.1
joblib==0.13.2
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.0.1
lazy-object-proxy==1.4.1
linear-regression==0.1
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mccabe==0.6.1
mistune==0.8.4
mpmath==1.1.0
nbconvert==5.4.1
nbformat==4.4.0
nltk==3.4.1
notebook==5.7.8
numpy==1.18.1
oauthlib==3.1.0
opencv-python==4.1.0.25
opt-einsum==3.1.0
pandas==0.24.2
pandocfilters==1.4.2
parso==0.4.0
pickleshare==0.7.5
Pillow==6.0.0
plotly==3.9.0
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.7
pydotplus==2.0.2
Pygments==2.3.1
pylint==2.3.1
pyparsing==2.4.0
pyrsistent==0.14.11
python-dateutil==2.8.0
python-weka-wrapper3==0.1.12
pytz==2019.1
pywinpty==0.5.5
PyYAML==5.1.2
pyzmq==18.0.1
qtconsole==4.4.3
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.0
Sastrawi==1.0.1
scikit-learn==0.20.3
scikit-plot==0.3.7
scipy==1.4.1
seaborn==0.9.0
Send2Trash==1.5.0
six==1.13.0
sklearn==0.0
soupsieve==1.9.5
sqlparse==0.3.0
sympy==1.5
tensorboard==2.1.0
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
textblob==0.15.3
tornado==6.0.2
traitlets==4.3.2
typed-ast==1.3.5
urllib3==1.25.7
virtualenv==16.7.9
virtualenvwrapper-win==1.2.5
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.4.2
wrapt==1.11.2
xlrd==1.2.0
xlwt==1.3.0
